requirements/base.txt
# Django
Django==3.1.3

# Django REST Framework
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0

# Hashers
argon2-cffi==20.1.0


# Base
Pillow==8.0.1

# Documentation
coreapi==2.3.3

# Environment
django-environ==0.4.5

mysqlclient==1.4.3